% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrativeATE.r
\name{IntegrativeATE}
\alias{IntegrativeATE}
\title{Integrative Average Treatment Effect (IntegrativeATE)}
\usage{
IntegrativeATE(I, x, u, y, A, method_val, method_ep, nboot)
}
\arguments{
\item{I}{the vector of the binary indicator of the validation sample membership; i.e., 1 if the unit belongs to the validation sample, and 0 otherwise  (n x 1)}

\item{x}{the matrix of confounders fully observed  (n x dim(x))}

\item{u}{the matrix of confounders observed only for the validation sample; and NA otherwise (n x dim(u))}

\item{y}{the vector of outcome (n x 1)}

\item{A}{the vector of binary treatment (n x 1)}

\item{method_val}{the estimation method for ATE on the validation sample

select 1 method from c("reg","ipw","aipw","matching")

\code{"reg"}: a linear regression imputation estimator of the ATE

\code{"ipw"}: the inverse probability of treatment weighting estimator of the ATE, where the propensity score follows a logistic regression model

\code{"aipw"}: the augmented inverse probability of treatment weighting estimator (Lunceford Davidian, 2004) of the ATE

\code{"matching"}: the matching estimator (Abadie and Imbens, 2006) of the ATE, using matching based on (X,U) with replacement with the number of matches fixed at M.}

\item{method_ep}{the error prone estimation method applied to the validation sample and the main sample

select >=1 methods from c("none","reg","ipw","aipw","matching")

\code{"none"}: return the initial estimator based soly on the validation sample}

\item{nboot}{the number of bootstrap samples; if nboot=0, then return only the variance estimator based on the asymptotic result}
}
\value{
\itemize{

\item \code{est}: estimate of the ATE

\item \code{ve}:  variance estimate for \code{est} based on the asymptotic result

\item \code{ve_boot}: variance estimate for \code{est} based on wild bootstrap
}
}
\description{
Implements integrative analyses for the average treatment effect combining main data with unmeasured confounder
and validation data with supplementary information on these confounders.
}
\details{
Under the unconfoundedness assumption
with completely observed confounders, the smaller validation data allow for constructing consistent estimators
for causal effects, but the big main data can only give error-prone estimators in general.

The integrative estimator leverages the information in the big main data to improve the estimation
efficiencies yet preserve the consistencies of the initial estimators based solely on the validation data.
Specifically, it uses the difference of the error-prone estimators applied to the main data and the validation data,
which is consistent for zero assuming that the main data and the validation data are representative of the same target population.

The framework applies to asymptotically normal estimators, including the commonly-used regression imputation,
weighting, and matching estimators.
}
\examples{

n<-1000  # the (combined) main sample size (Samples 1 and 2)
n2<-500  # the validation sample size (Sample 2)

## generate covariates (x,u)

x<-runif(n,0,2)
u<-2*sin(x)+runif(n,0,1)

## generate treatment A

lps<-cbind(1,x,u)\%*\%c(0,-1,1)
ps<-exp(lps)/(1+exp(lps))
A<-rbinom(n,1,ps)

## generate potential and observed outcomes

loc.a1<-which(A==1)
loc.a0<-which(A==0)
y1<- -x+2*u+rnorm(n,0,1)
y0<- -x+1*u+rnorm(n,0,1)
y<-y1*A+y0*(1-A)
true<-mean(y1-y0)

## generate indicator of membership of the validation sample (Sample 2)

I<-rep(1,n)
I[((1+(n-n2)):n)]<-0

## u is not observed for Sample 1
loc1<-which(I==1)
u[loc1]<-NA


method_val <-c("reg")
method_ep<-c("reg","ipw")

true
out<-IntegrativeCI::IntegrativeATE(I,x,u,y,A,method_val,method_ep,nboot=50)
out$est
out$ve
out$ve_boot


}
\references{
Yang, S. and Ding, P. (2018). Combining multiple observational data sources to estimate causal effects.
\url{https://arxiv.org/abs/1801.00802}
}
